JWT Co-authored-by: luchomellu <luchomellu@users.noreply.github.com> Co-authored-by: Franco Perez <FrancoGabrielPerez@users.noreply.github.com>

+++ tpFinal/microAdministration/src/main/java/com/microadministration/controller/AdminStaffController.java
@@ -47,7 +47,12 @@ public class AdminStaffController {
      */
     @Operation(summary = "Obtiene todos los integrantes del staff de administracion.", description = "Obtiene todos los integrantes del staff de administracion.")
     @GetMapping("")
-    public ResponseEntity<?> getAll(){
+    public ResponseEntity<?> getAll(@RequestHeader("Authorization") String token){
+        ResponseEntity<String> response = validarToken(token);
+
+        if (response.getStatusCode() != HttpStatus.OK) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Token no válido");
+        }
         try{
             return ResponseEntity.status(HttpStatus.OK).body(adminService.findAll());
         }catch (Exception e){
@@ -62,7 +67,12 @@ public class AdminStaffController {
      * @return ResponseEntity<?>
      */
     @GetMapping("/buscar/rol/{rol}")
-    public ResponseEntity<?> getByRol(@PathVariable String rol){
+    public ResponseEntity<?> getByRol(@RequestHeader("Authorization") String token, @PathVariable String rol){
+        ResponseEntity<String> response = validarToken(token);
+
+        if (response.getStatusCode() != HttpStatus.OK) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Token no válido");
+        }
         try{
             return ResponseEntity.status(HttpStatus.OK).body(adminService.findByRol(rol));
         }catch (Exception e){
@@ -77,7 +87,12 @@ public class AdminStaffController {
      */
     @Operation(summary = "Agrega un nuevo intregrante al staff.", description = "Agrega un nuevo intregrante al staff")
     @PostMapping("/alta")
-    public ResponseEntity<?> save(@RequestBody AdminStaffDTO entity){
+    public ResponseEntity<?> save(@RequestHeader("Authorization") String token, @RequestBody AdminStaffDTO entity){
+        ResponseEntity<String> response = validarToken(token);
+
+        if (response.getStatusCode() != HttpStatus.OK) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Token no válido");
+        }
         try{            
             return ResponseEntity.status(HttpStatus.OK).body(adminService.save(entity));
         }catch (Exception e){
@@ -93,7 +108,12 @@ public class AdminStaffController {
      */
     @Operation(summary = "Obtiene un integrante por su id.", description = "Obtiene un integrante del staff por su staffId")
     @GetMapping("/buscar/id/{staffId}")
-    public ResponseEntity<?> getById(@PathVariable long staffId) {
+    public ResponseEntity<?> getById(@RequestHeader("Authorization") String token, @PathVariable long staffId) {
+        ResponseEntity<String> response = validarToken(token);
+
+        if (response.getStatusCode() != HttpStatus.OK) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Token no válido");
+        }
         try{
             return ResponseEntity.status(HttpStatus.OK).body(adminService.findById(staffId));
         }catch (Exception e){
@@ -109,7 +129,12 @@ public class AdminStaffController {
      */
     @Operation(summary = "Elimina un integrante del staff por su id.", description = "Elimina un integrante del staff por su Id")
     @DeleteMapping("/eliminar/{staffId}")
-    public ResponseEntity<?> delete(@PathVariable long staffId){
+    public ResponseEntity<?> delete(@RequestHeader("Authorization") String token, @PathVariable long staffId){
+        ResponseEntity<String> response = validarToken(token);
+
+        if (response.getStatusCode() != HttpStatus.OK) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Token no válido");
+        }
         try{
             adminService.delete(staffId);
             return ResponseEntity.status(HttpStatus.OK).body("Se elimino correctamente al integrante con Id: " + staffId);
@@ -127,7 +152,13 @@ public class AdminStaffController {
      */
     @Operation(summary = "Actualiza los datos de un integrante del staff por su id.", description = "Actualiza un integrante del staff por su id")
     @PutMapping("/actualizar/{staffId}")
-    public ResponseEntity<?> update(@PathVariable long staffId, @RequestBody AdminStaffDTO entity){
+    public ResponseEntity<?> update(@RequestHeader("Authorization") String token, @PathVariable long staffId, @RequestBody AdminStaffDTO entity){
+        ResponseEntity<String> response = validarToken(token);
+
+        if (response.getStatusCode() != HttpStatus.OK) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Token no válido");
+        }
+
         try{
             adminService.update(staffId, entity);
             return ResponseEntity.status(HttpStatus.OK).body("Se actualizo correctamente al usuario con Id: " + staffId);
